More Preparations.  Release probably soon
Redid the Readme to talk about Void Engine instead of Kade, although it does mention that this was Kade Engine Originally.  Changed the in-game indicator that had the version to say Void Engine with the new version number.  This next update includes a couple of fixes from Kade, as well as unhardcoded weeks allowing modCore to utilize it.